reasoning/requirements.txt
# Phase 5 retrieval layer + Phase 6+ reasoning engine
llama-index-graph-stores-neo4j>=0.3.0
llama-index-vector-stores-qdrant>=0.3.0
llama-index-embeddings-fastembed>=0.2.0
llama-index-core>=0.11.0
fastembed>=0.3.0
qdrant-client>=1.14.0,<1.17.0
neo4j>=5.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
